+++ CHANGELOG.md
@@ -5,6 +5,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- ## [Unreleased] -->
 
+## [0.3.1] - 2022-05-03
+
+### Added
+- Option to force mixdown in every case
+- range optimisations
+
 ## [0.3.0] - 2022-05-02
 
 ### Added